feat: add Pickle trait
This allows the application to control serialization. Doing this can
allow the application to choose the serialization format they want. It
can also allow the application to choose whether to derive serialize
on their aggregate or not. For example they may want a mapping to some
other type to prevent serializing their aggregate and making it more
annoying to change.